Removal of _new Suffix from database tables and related files
