--- server/src/models/SessionNew.js
@@ -1,29 +1,0 @@
-class SessionNew {
-  constructor(
-    sessionId,
-    deviceId,
-    createdAt,
-    updatedAt,
-    sessionStatus,
-    attendingResponderNumber,
-    doorOpened,
-    surveySent,
-    selectedSurveyCategory,
-    responseTime,
-    sessionRespondedVia,
-  ) {
-    this.sessionId = sessionId
-    this.deviceId = deviceId
-    this.createdAt = createdAt
-    this.updatedAt = updatedAt
-    this.sessionStatus = sessionStatus
-    this.attendingResponderNumber = attendingResponderNumber
-    this.doorOpened = doorOpened
-    this.surveySent = surveySent
-    this.selectedSurveyCategory = selectedSurveyCategory
-    this.responseTime = responseTime
-    this.sessionRespondedVia = sessionRespondedVia
-  }
-}
-
-module.exports = SessionNew
